+++ controls/controlsCursor.js
@@ -91,7 +91,7 @@ function cursor_draw(){
 
 function cursor_set_shape(){
 		//Clear cursor chunk array and clear cursot list to populate with new values
-		controls.cursorChunkData = new Uint8Array(Math.pow(blockSettings.chunk.XYZ,3)).fill(255);
+		controls.cursorChunkData = new Float32Array(Math.pow(blockSettings.chunk.XYZ,3)).fill(255);
 		controls.cursorList = [];
 		switch(controls.cursorShape){
 			//None
